# 64-bit integer arithmetic performance optimization