--- a/compiler/natives/src/math/big/big_test.go
+++ b/compiler/natives/src/math/big/big_test.go
@@ -12,3 +12,7 @@ func TestBytes(t *testing.T) {
 func TestModSqrt(t *testing.T) {
 	t.Skip("slow")
 }
+
+func TestLinkerGC(t *testing.T) {
+	t.Skip("The test is specific to GC's linker.")
+}