--- a/tests/js_test.go
+++ b/tests/js_test.go
@@ -1,5 +1,5 @@
-//go:build js
-// +build js
+//go:build js && !wasm
+// +build js,!wasm
 
 package tests_test